+++ src/main/java/org/generation/fyndr/repositorios/ResenaRepository.java
@@ -6,5 +6,5 @@ import org.springframework.data.jpa.repository.JpaRepository;
 /**
  * Interfaz de repositorio para la entidad Resena.
  */
-public interface ResenaRepository extends JpaRepository<Resena, Integer> {
+public interface ResenaRepository extends JpaRepository<Resena, Long> {
 } // interface ResenaRepository
